+++ core/src/test/java/nl/weeaboo/vn/impl/scene/ScreenTest.java
@@ -1,16 +1,31 @@
 package nl.weeaboo.vn.impl.scene;
 
 import org.junit.Assert;
+import org.junit.Before;
 import org.junit.Test;
 
+import nl.weeaboo.vn.core.SkipMode;
+import nl.weeaboo.vn.impl.core.SkipState;
+import nl.weeaboo.vn.impl.core.TestEnvironment;
 import nl.weeaboo.vn.impl.test.CoreTestUtil;
+import nl.weeaboo.vn.input.VKey;
 import nl.weeaboo.vn.scene.ILayer;
 
 public class ScreenTest {
 
+    private TestEnvironment env;
+    private SkipState skipState;
+    private Screen screen;
+
+    @Before
+    public void before() {
+        env = TestEnvironment.newInstance();
+        skipState = new SkipState();
+        screen = CoreTestUtil.newScreen(skipState);
+    }
+
     @Test
     public void layers() {
-        Screen screen = CoreTestUtil.newScreen();
         ILayer active = screen.getActiveLayer();
         Assert.assertNotNull(active); // Active layer should never be null
         ILayer root = screen.getRootLayer();
@@ -22,6 +37,18 @@ public class ScreenTest {
         Assert.assertTrue(root.containsLayer(subLayer));
         Assert.assertTrue(root.containsLayer(subSubLayer));
         Assert.assertTrue(subLayer.containsLayer(subSubLayer));
+    }
+
+    @Test
+    public void testAutoReadCancel() {
+        skipState.setSkipMode(SkipMode.AUTO_READ);
+        Assert.assertEquals(SkipMode.AUTO_READ, skipState.getSkipMode());
+
+        // Pressing the text continue key cancels auto read mode
+        env.getInput().buttonPressed(VKey.TEXT_CONTINUE);
+        screen.update();
+
+        Assert.assertEquals(SkipMode.NONE, skipState.getSkipMode());
     }
 
 }